CampusCompanionAgent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.25
langchain-openai==0.3.18
langchain-community==0.3.21
httpx>=0.27.0
httpx-sse>=0.4.0
python-dotenv==1.0.1
pydantic==2.9.2
sse-starlette==2.1.3
